requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.3.4
async-timeout==4.0.2
attrs==21.4.0
bcrypt==4.0.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==38.0.1
defusedxml==0.7.1
Django==3.2
django-allauth==0.51.0
django-appconf==1.0.5
django-auto-logout==0.5.0
django-bootstrap-modal-forms==2.2.0
django-ckeditor==6.5.0
django-cleanup==6.0.0
django-downloadview==2.3.0
django-extensions==3.2.0
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-model-utils==4.2.0
django-pagination==1.0.7
django-users2==0.2.2
django-utils-six==2.0
frozenlist==1.3.0
graphviz==0.20.1
idna==3.3
multidict==6.0.2
mysqlclient==2.1.1
oauthlib==3.2.1
Pillow==8.4.0
psycopg2==2.9.1
pycparser==2.21
pydot==1.4.2
PyJWT==2.4.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2021.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.12
yarl==1.7.2
